requirements.txt
torch
transformers
numpy
pandas
pillow
streamlit

